Rename bfxapi._utils.logger to bfxapi._utils.logging (and update references).
